+++ commons/src/main/java/commons/GuessQuestion.java
@@ -33,7 +33,7 @@ public class GuessQuestion extends Question{
      * @return Closeness of the guess to the correct answer.
      */
     public double calculateCloseness(int guess) {
-        int correctAnswer = this.getActivity().getConsumption_in_wh();
+        long correctAnswer = this.getActivity().getConsumption_in_wh();
         // Formula: if the guess is 100% away from the actual answer, the closeness should be 0. Else, it is 1-error
 
         if (guess >= 2*correctAnswer || guess <= 0) {
